requirements.txt
# Web Framework
Flask==3.1.3
Werkzeug==3.1.6
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

# Machine Learning & Data Science
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
scikit-learn==1.7.2
scipy==1.15.3
pandas==2.3.3
numpy==2.2.6

# Visualization & Image Processing
matplotlib==3.10.8
seaborn==0.13.2
Pillow==12.1.1

# Utilities
requests==2.32.5
tqdm==4.67.3
kaggle==1.7.4.5
colorama==0.4.6
markupsafe==3.0.3